+++ packages/necto-platform/src/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Copyright (c) Corinvo, LLC. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ */
+
+// This is a horrible device detector code, DO NOT USE IN ANY PROD CODE!!!
+
+export * from "./isMac";
+export * from "./isAndroid";