+++ packages/necto-platform/src/isAndroid.ts
@@ -1,0 +1,14 @@
+function testUserAgent(re: RegExp) {
+  if (typeof window === 'undefined' || window.navigator == null) {
+    return false;
+  }
+  return (
+    //@ts-ignore
+    window.navigator['userAgentData']?.brands.some((brand: {brand: string, version: string}) => re.test(brand.brand))
+  ) ||
+  re.test(window.navigator.userAgent);
+}
+
+export const isAndroid = function () {
+  return testUserAgent(/Android/i);
+};