+++ packages/necto-react-helpers/src/isKeyboardFocusEvent.ts
@@ -1,4 +1,15 @@
-'use strict';
+/**
+ * Copyright (c) Corinvo, LLC. and affiliates.
+ *
+ * This source code is licensed under the MIT license found in the
+ * LICENSE file in the root directory of this source tree.
+ *
+ * Portions of this code are based on the React Aria Spectrum library by Adobe,
+ * licensed under the Apache License, Version 2.0.
+ * See: https://github.com/adobe/react-spectrum
+ *
+ * Modifications have been made to adapt the code for use in this project.
+ */
 
 import { getOwnerDocument, getOwnerWindow } from "@necto/dom";