+++ packages/necto-react-hooks/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @necto-react/hooks
 
+## 2.1.0
+
+### Minor Changes
+
+- Added new hooks and a new platforms package
+
+### Patch Changes
+
+- Updated dependencies
+  - @necto-react/helpers@2.1.0
+  - @necto/platform@1.1.0
+
 ## 2.0.0
 
 ### Major Changes